mkinitramfs: Add lib search path + run full ldconfig in initramfs
Allmost directly taken from dracut.
Do it also at a late stage when everything is on initramfs.

It should fix boot on a box where the lvm library was searching for a
libgcc, which was on the initramfs but not in the lib search path.

Ubuntu got hit similarly:
"We had a rather obscure Ubuntu bug
(https://bugs.launchpad.net/bugs/728611) which presented as Plymouth not
being able to display text at boot time, which of course is nasty in
cases where the user needs to answer a prompt in order for boot to
complete.  This turned out to be restricted to configurations where
Plymouth is built into the initramfs, so I chased this down.

The problem is that /lib/plymouth/label.so now transitively links to
libGL, which (on Ubuntu, at least) lives in /usr/lib/mesa/libGL.so.1 or
some other subdirectory of /usr/lib depending on the GL implementation
in use (fglrx, nvidia, etc.).  This directory isn't on the linker's
default search path; instead, it relies on /etc/ld.so.cache having been
built appropriately from a configuration including
/etc/ld.so.conf.d/GL.conf, which is managed by update-alternatives.  The
thing that's going wrong here is that /usr/lib/mesa/libGL.so.1 is copied
into the initramfs, but it isn't on the linker's search path so
/lib/plymouth/label.so fails to load."  -cjwatson

Closes: #612633, #619670
Reported-by: Benedikt Spranger <b.spranger@linutronix.de>
Tested-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: maximilian attems <max@stro.at>
Reported-by: Colin Watson <cjwatson@ubuntu.com>
[ code readability fixes -mika ]
Signed-off-by: Michael Prokop <mika@debian.org>