mkinitramfs: copy over on build instead of using symlink tree
This facilitates mkinitramfs building.
It is intended as fix for the not possible ldconfig run inside
the symlink tree. It should also allow to cleanup the copy_exec() mess.

This also allows to no longer shipp duplicate busybox copies in
initramfs. dracut does it also. The symlinking was thought as
speed gain by jbailey, but I'm less then conviced it is worth it.

Also specify to cpio that the files should be owned by root.

In retrospect this really similar quasi a fowarded ported patch by
 Michal Schmidt <xschmi00@stud.feec.vutbr.cz>

Closes: #338405, #506540
Signed-off-by: maximilian attems <max@stro.at>
Tested-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Michael Prokop <mika@debian.org>