debian/copyright: Update authorlist and year attribution
List the top 6 commiters of the initramfs-tools git repo.
Update the year attribution, also pay attention to current maintainers.
No point in listing people that didn't contribute really in a
significant way.

The cut off of the best 6 is quite arbitrary. It just happens to be
that those have each more then 5 commits / person and thus can be
seen as significant contributors.

Signed-off-by: maximilian attems <max@stro.at>
Reviewed-by: Michael Prokop <mika@debian.org>