initramfs-tools: init mount /run tmpfs
The purpose is to have a persisent directory from boot on.

Also migrate our debug log to /run:
Use /run/initramfs and not /dev/.initramfs

This was a private dir with no good outside usage
(usplash is gone, /etc/init.d/rc mentions progress bar).
live-initramfs needs to use /run/initramfs for uninofs-fuse.

If the rootmnt has no /run dir care to umount the dir in order
to not confront user with mount move errors.
This leeds to debug info loosage, which is considered ok.

Signed-off-by: maximilian attems <max@stro.at>
Reviewed-by: Christian Hofstaedtler <ch@grml.org>
Reviewed-by: Michael Prokop <mika@debian.org>