init: sync mount options with the defaults in initscripts
This bit syncs the mount options with the initscripts defaults, but
most importantly the s/none/tmpfs/ makes the initscripts mount checks
not mount a second tmpfs on /run, obscuring the one from the initramfs.
domount() checks for already mounted filesystems to guard against
mounting an already mounted filesystem.  But it mounts all tmpfs
filesystems with "tmpfs" as mnt_fsname, and having this as "none"
breaks that check, leading to a second tmpfs mount.  Please could you
do the s/none/tmpfs whether or not you also pick up the mount defaults
changes, which are optional (we remount with the user-defined options
later).

Note that the nodev option was removed, because it's currently
required on some systems by /etc/init.d/checkroot.sh (it uses
mknod /dev/rootdev to get a device node to mount /root).  I hope
that we can remove this at some point, but for now mounting /run
with nodev will cause some systems to break.  It also mounts with
noexec, since this is potentially needed for some uses of /dev/shm.
Again, I hope we can remove it, but it will need testing whether
or not exec perms are needed to map with PROT_EXEC on current
kernels.

WRT these mount options, I aim to get initscripts to remount /run
with options from /etc/fstab or /etc/default/tmpfs which will
allow the system to be configured to use more restrictive mount
options after boot, should the admin choose.  And, depending upon
the configuration, we may be able to automatically determine if
stricter permissions are possible (e.g. /run/shm is a separate
mount, so noexec is safe to use).  So this patch uses the options
we know are safe, and initscripts can tighten them up later, if
possible.  This remounting with user-specified options will apply
to all filesystems mounted in mountkernfs/mountdevsubfs.

Closes: #621803

Signed-off-by: rleigh <rleigh@codelibre.net>
Tested-by: Tollef Fog Heen <tfheen@err.no>
Tested-by: Michael Biebl <biebl@debian.org>
Signed-off-by: maximilian attems <max@stro.at>
Reviewed-by: Christian Hofstaedtler <ch@grml.org>
Reviewed-by: Michael Prokop <mika@debian.org>