init: Preserve state on old boxes without /run on rootfs
If /run is not present on the rootfs, instead of losing the /run/udev from the
initramfs, we move it to /dev/.udev:

This is against initramfs-tools branch maks/run.  This preserves historical
behaviour, and will make udev/lvm work correctly during the transitional
period, i.e. the state is kept in the same place it always has been.  Once
you install a new initscripts which puts a /run on the rootfs, the udev
state will stay in /run.  So this caters for either possibility.

Signed-off-by: rleigh <rleigh@codelibre.net>
[ shorten comments -maks ]
Signed-off-by: maximilian attems <max@stro.at>
Reviewed-by: Christian Hofstaedtler <ch@grml.org>
Reviewed-by: Michael Prokop <mika@debian.org>