init: Don't use "none" as a mount fsname
When mounting kernel filesystems and tmpfs filesystems, "none" is
unsuitable as the filesystem name.  This is because it causes
breakage when /etc/mtab is a symlink to /proc/mounts.  This is needed
so that we can switch /etc/mtab to be a symlink when util-linux is
updated to use libmount (#620710).

When /etc/mtab is a file updated by mount, the initscripts
mountkernfs.sh, mountdevsubfs.sh and mtab.sh initialise the mtab at
boot using values other than "none".  For example, /proc uses "proc",
/dev/pts uses "devpts" and /sys uses "sysfs".  The mtab entries are
created from defaults, plus any user customisations added to
/etc/fstab.

When /etc/mtab is a symlink to /proc/mounts, the values set when the
filesystems were mounted by init in the initramfs are now visible, and
these differ from values in /etc/fstab.  The initscripts are now no
longer able to update the mtab file, and this causes problems mounting
local filesystems.

The most visible issue is /proc.  The default fstab entry is:
  proc            /proc           proc    defaults        0       0
while init is using
  mount -t proc -o nodev,noexec,nosuid none /proc
rather than
  mount -t proc -o nodev,noexec,nosuid proc /proc
which results in "mount -a" failing with an exit code of 32 (mount
failure).  While altering "proc" to "none" in the fstab file corrects
this boot failure, this is present by default in the fstab file of
every installed system, and so is not easy to update.  Additionally,
the /dev/pts and /sys mounts may also have been added using the
"devpts" and "sysfs" names, and so these should ideally also be
corrected to use these names in the initramfs.  This behaviour may be
easily verified by changing the fsname for /proc in /etc/fstab and
checking the exit status of "mount -a".

I also fixed an inconsistency in the udev /dev fsname.  It might use
"none" or "udev"; it now always uses "udev".

Closes: #603858
Reported-by: Goswin von Brederlow <goswin-v-b@web.de>
Signed-off-by: Roger Leigh <rleigh@debian.org>
[ fixes bug in old initscripts barking on 'none' fsname. -maks ]
Signed-off-by: maximilian attems <max@stro.at>
Reviewed-by: Christian Hofstaedtler <ch@grml.org>
Reviewed-by: Michael Prokop <mika@debian.org>