mark Multi-Arch: foreign
It would be useful for initramfs-tools to be marked "Multi-Arch:
foreign", to indicate that it can satisfy dependencies of packages of
other architectures.  Although Debian's dpkg doesn't yet do anything
special with this, it's safe to add this tag in advance of package
manager support.

In Ubuntu, this is an early step in being able to crossgrade from i386
to amd64, because we don't have an -amd64 kernel flavour on i386 and (of
course) our kernel packages depend on initramfs-tools.  While Debian
does have an -amd64 flavour on i386, it still wouldn't hurt to add this
tag.

If you're wondering why this tag is needed on an Architecture: all
package, see the multiarch specification:

  https://wiki.ubuntu.com/MultiarchSpec#Dependencies_involving_Architecture:_all_packages

Closes: 649399
Signed-off-by: Michael Prokop <mika@debian.org>
Signed-off-by: maximilian attems <max@stro.at>