initramfs-tools: copy_exec add pattern for nonoptimized multiarch libraries
initramfs-tools hook-functions has copy_exec function needs some
updates.

There is one unquoted occurence of ${src} and the pattern for
nonoptimized libraries does not match multiarch libraries.

It does not match multilib libraries in /lib32 and /lib64 but those are
rare and should not multiply. The only library that is installed in
/lib32/i586 and /lib32/i686/cmov is libssl on mu system.

Note that for x in ... does not handle spaces in library paths but I am
not particularly concerned.

Closes: #659752
Signed-off-by: maximilian attems <max@stro.at>