init: fixes ignored $tmpfs_size
* What led up to the situation?
The mouted size of /dev is larger then 10M (which is default, in the init
script itself).

* What exactly did you do (or not do) that was effective (or ineffective)?
On examining the /usr/share/initramfs-tools/init I found a skipp over or typo.

* What was the outcome of this action?
I patched the init script, rebuild/updated /boot/initrd-<whatever> using
update-initramfs -u; rebooted the system and report size now is 10M :)

* What outcome did you expect instead?
The udevfs/tmfs size for the /dev should be 10M (default) or as per
/etc/udev/udev.conf

Closes: #670496
Signed-off-by: maximilian attems <max@stro.at>