hook-functions: handle rootfs output from mount(8)
During my upgrade, sysvinit-utils and initscripts have been upgraded
from 2.88dsf-13.1 to 2.88dsf-22.1. I think (I did not check) that this
move /etc/mtab into a symlink to /proc/mounts. So, the result of "mount"
should be handled the same as the contents of /proc/mounts.  And indeed,
adding " && [ "$fs" != "rootfs" ]" when handling the output of mount
fixed the problem for me.

Closes: #668616
Reported-by: Vincent Danjean <vdanjean@debian.org>
Signed-off-by: maximilian attems <max@stro.at>