init: tighten /run/initramfs permissions
When the system is booted and users can log in, there is nothing
to stop a user denial of service by filling up /run through the
creation of files in /run/initramfs.  I can't think of any valid
reason to give a user write access to a filesystem only intended
to be writable by system processes.

I would suggest creating it with 0755 permissions for safety and
security.

Closes: #655175
Reported-by: Roger Leigh <rleigh@debian.org>
Signed-off-by: maximilian attems <max@stro.at>