initramfs-tools: fix update-initramfs with modprobe options
Ignore module options printed by modprobe, long story:

> Using update-initramfs -u gives me the following output:
>
> # update-initramfs -u
> update-initramfs: Generating /boot/initrd.img-3.3.0-trunk-amd64
> install: cannot stat `/lib/modules/3.3.0-trunk-amd64/kernel/drivers/hid/hid-apple.ko fnmode=2 ': No such file or directory
> ERROR: Module /lib/modules/3.3.0-trunk-amd64/kernel/drivers/hid/hid-apple.ko fnmode=2  not found.
> install: cannot stat `/lib/modules/3.3.0-trunk-amd64/kernel/drivers/gpu/drm/radeon/radeon.ko modeset=1 ': No such file or directory
> E: /usr/share/initramfs-tools/hooks/plymouth failed with return 1.
> update-initramfs: failed for /boot/initrd.img-3.3.0-trunk-amd64 with 1.
>
> # cat /etc/modprobe.d/radeon-kms.conf
> options radeon modeset=1
>
> # dpkg -S /etc/modprobe.d/radeon-kms.conf
> xserver-xorg-video-radeon: /etc/modprobe.d/radeon-kms.conf

Looks like modprobe's current behavior doesn't quite match what it
appeared to do when I originally wrote the patch for bug 652525; in
particular, modprobe --show-depends now outputs the options for a
module, not just the module.  (It must not have done so at the time I
wrote the patch, or I would have seen this failure when I tested the
patch on my system, since I have the same option set for i915.)


When reading the output of modprobe, only use the first word after
"insmod" as the module name, not the remainder of the line (which may
include module options).

Closes: #676400
Signed-off-by: Josh Triplett <josh@joshtriplett.org>
Signed-off-by: maximilian attems <max@stro.at>