Use setupcon to install system's keymap
/etc/console-setup/cached.kmap.gz doesn't necessarily exist any
longer, so we shouldn't hardcode the path but instead rely on
setupcon's --save-keyboard option to install the proper keymap.

The --save-keyboard option was introduced as of console-setup v1.72.
There add add Breaks: console-setup (<< 1.72) as we rely on the
--save-keyboard option.

Thanks: Andreas Altergott <andreasaltergott@googlemail.com> for initial patch
Thanks: Raphaël Hertzog <hertzog@debian.org> for testing, feedback and review
Closes: #619711