keymap hook: provide warning message if loadkeys/setupcon are not available
When a disk needs to be decrypted in the initramfs stage then
loadkeys (provided by kbd) and setupcon (provided by
console-setup) should be available.  The cryptsetup hook script
/usr/share/initramfs-tools/conf-hooks.d/cryptsetup sets KEYMAP=y,
so if loadkeys/setupcon are not available then let's inform the
user about it, as initramfs-tools version 0.108 requires both
loadkeys and setupcon. See also #689722.

Thanks: Raphael Hertzog <hertzog@debian.org> for the feedback
Closes: #689336