hook-functions: Include more HID drivers in base
Currently we're not including all the HID drivers that support
keyboards.  Further, there are occasional regressions when a keyboard
that was previously handled by the generic code gets a specialised
drivers that's not on our list.

Instead of trying to list all the HID driver modules, include the
whole drivers/hid tree and exclude specific modules that we know don't
support keyboards.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Closes: #697619