resolve_device: Canonicalise symbolic links to real device nodes
Canonicalise device names to match util-linux mount behaviour.
This ensures that "mount -a" in mountall does not try to mount
/usr a second time (which it will attempt if the mounted device
does not match the canonical device name).  This also fixes a
longstanding annoyance with the output of mount and df using long
UUIDs rather than short device names.

Note that resolve_device may be called more than one; it's done
during options parsing for ROOT, and again immediately
prior to mounting.  This is because during options parsing the
device node does not exist, but we still need to construct a
path into /dev from the LABEL or UUID.

Rebased and ETC handling dropped by Michael Prokop <mika@debian.org>