lsinitramfs(8): document failure to deal with early initramfs
lsinitramfs cannot deal with a multi-segmented initramfs archive, such
as one with an uncompressed early initramfs prepended to the main
compressed initramfs.  Document this in the manpage.

The kernel will parse all cpio archives it can find in the initramfs in
sequence, and it doesn't care if some of them are compressed and others
are not.

Signed-off-by: Henrique de Moraes Holschuh <hmh@debian.org>