hooks: Add resume hook instead of hardcoding RESUME once on preinst
This logic is better run every time on update-initramfs,
as swap partition of a system might change.
Also there are scenarios where the preinstall picks up
a wrong value that stays wrongly hardcoded.
This results in a 5s useless wait on boot.

The biggest swap partition is used as valid guess.
This was previously the logic and is the logic used by ubiquity
(Ubuntu live installer) too.

Closes: #565225, LP 50437.

Thanks-to: Dmitrijs Ledkovs <launchpad@surgut.co.uk>
Reviewed-by: Martin Pitt <martin.pitt@ubuntu.com>
Reviewed-by: Michael Prokop <mika@debian.org>
Signed-off-by: maximilian attems <maks@debian.org>