hooks/function: More solide ip= option parsing
The configure_networking() function in scripts/functions fails
to parse an "ip=" kernel option if it has more than one parameter
following the device name.  For example:

   ip=client-ip:server-ip:gw-ip:netmask:hostname:device:autoconf

works just fine, however,

   ip=client-ip:server-ip:gw-ip:netmask:hostname:device:autoconf:dns0:dns1

does not, and that happens to be the format described in
   https://www.kernel.org/doc/Documentation/filesystems/nfs/nfsroot.txt

A one character change to configure_networking() will make it immune to
any number of parameters following the device name.

Closes: #724644

Reviewed-by: Michael Prokop <mika@debian.org>
Signed-off-by: maximilian attems <maks@debian.org>