debian/control: Add Breaks for console-setup (>= 1.71~)
console-setup 1.71 got rid of cached.kmap.gz (Debian bug #619711) so
we had to change keymap loading in initramfs-tools 0.108.
Unfortunately the new implementation is not compatible with older
versions of console-setup, as in squeeze.

I don't feel like solving the problem of compatibility with both old
and new console-setup versions, so I have reverted to the old
implementation in this backport.  Add Breaks to ensure that a
squeeze+backports -> wheezy upgrade will upgrade initramfs-tools
together with console-setup.  (Neither squeeze's initramfs-tools nor
wheezy's console-setup has this!)

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>