get_fstype: initialize FSTYPE variable
On the hppa platform the fstype program from klibc crashed (details in Bug#745660) which exposed a problem in the
get_fstype() shell function in file scripts/functions.

In this script, fstype is called like this:
        eval $(fstype "${FS}" 2> /dev/null)
        if [ "$FSTYPE" = "unknown" ] ....
Since fstype crashed and returned nothing in the variable "FSTYPE", FSTYPE stayed empty instead of the value "unknown" and as such no further analysis via blkid was done.

I think it makes sense to pre-initialize FSTYPE to the value "unknown" before
calling fstype. That way the program logic will continue correctly if something
with the fstype program is wrong.

Closes: #745731

Signed-off-by: Helge Deller <deller@gmx.de>
Signed-off-by: maximilian attems <maks@debian.org>