hook-functions: add support for virtio-mmio
On most virtual machines it is possible to use the virtio interface
through the PCI bus to export for example block or net devices. On some
architectures the emulated machine does not have a PCI bus, and thus the
transport goes through an MMIO interface.

Currently initramfs-tools correctly handle the PCI case, but not the
MMIO on. In case of a virtio based root device, the virtio_mmio module
is not available in the initramfs causing the boot to fail. This commit
adds the virtio_mmio module if virtio is used (in dep mode), or by
default (in most mode).

Closes: #751143
Signed-off-by: Michael Prokop <mika@debian.org>