manual_add_modules: Set the kernel version when checking module firmware deps
Currently we don't tell modinfo which kernel version to look at
when checking which firmware files a module may need.  It will
default to the running kernel, not the one we're building the
initramfs for.  So we may not copy all the necessary firmware
or we may provoke warnings in case a module doesn't exist in
the running kernel.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>