Check for root device's module dependencies at all levels of device hierarchy
Currently sys_walk_modalias only looks for a modalias in the
grandparent of the given device path.  This doesn't match what the
name 'walk' implies.

sys_walk_mod_add calls it for each level of the device hierarchy, but
since it only looks at the grandparent it will miss some modaliases,
e.g. sd_mod or virtio_blk.  We explicitly add sd_mod when we see any
SCSI device, so we usually get away with this.  However, virtio_blk is
currently not added to a MODULES=dep initramfs if it is modular for
the target kernel but built into the running kernel.

Related-to: #760127
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>