panic: Do not return if reboot fails
The panic function may be called from several places in the init
script, not just at the end.  In some cases, the calling function will
retry an action when it returns; for example, booting with
'root=/dev/nonexistent panic=1' currently leads to an infinite series
of error messages.

In order to trigger a kernel panic reliably, we must use exit rather
than return.

Related-to: #751488
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>