checkfs: Continue unless fsck reports uncorrectable errors
The current checking of the fsck exit code is extremely fragile.  We
should never see the reboot flag (2) set as that is only used when
checking a mounted root filesystem, so don't bother checking for it.
So we should only need to check whether the uncorrectable error flag
(4) is set, and should continue otherwise.

This mitigates bug #766448.

Closes: #765614
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>