checkfs: Correct error messages and emergency shell in case of failure
The current error messages and some code are taken from initscripts,
which will fsck an already-mounted root filesystem and then must
reboot if it is modified.  Here we have not mounted the filesystem and
can continue if the fsck was successful, or retry after running an
emergency shell.

The emergency shell also doesn't work (sulogin crashes); we must use
the panic function for this instead.

log_warning_msg does not print to the console if the 'quiet' parameter
was used, so is not suitable for warning messages that must appear
unconditionally.  Pass a terse explanatory message to the panic
function.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>