local: Invoke local-block scripts from the loop in local_device_setup
Currently we invoke the local-top script once (and possibly
local-block once for /usr) and then poll for block devices to appear
if that wasn't sufficient.  This works for synchronously discovered
devices, devices stacked on top of them (using lvm2/mdadm/cryptsetup)
and asynchronously discovered devices like USB storage.  It doesn't
work for devices stacked on top of USB unless they are set up entirely
via udev hooks.

We need to invoke local-block again before each time we poll.
As the local-block scripts might not be that quick and we don't
have a good way to include their run-time in the timeout check,
wait 1 second rather than 0.1 second between polls.

Closes: #678696
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>