local: Use max(30, rootdelay) as timeout for block device to appear
The rootdelay kernel parameter is specified as a fixed delay before
attempting to mount the root filesystem, not a timeout.  In the
initramfs it is processed by udev's init-top script.  This is being
used to workaround bug #678696, but as most users will set the
shortest value that seems to work, it is fragile.  Using the same
value as a timeout makes our fix for #678696 fragile too.

We shouldn't really use $ROOTDELAY here at all, but as some users may
have a good reason for wanting a longer timeout and may depend on the
current behaviour, make our current default timeout of 30 seconds a
minimum.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>